backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

pdfplumber>=0.10.0,<0.11
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

openai>=1.12.0
python-dotenv>=1.0.0
